[AAE-5958] arefactored mocks
